Move treeview.js under app/javascript/packs
